backend/requirements.txt
# Web Server & API
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-multipart==0.0.22

# Data Science
scikit-learn==1.8.0
pandas==3.0.1
numpy==2.4.3
scipy==1.17.1
joblib==1.5.3

# Auth & Security
pydantic[email]==2.12.5
python-jose[cryptography]==3.5.0
# passlib[bcrypt]==1.7.4
pwdlib[bcrypt]
python-dotenv==1.2.2